Separation of jvm-specific code from SQLStringContextQueryTest
